moved query to the utils, improved the jikan anime search results
